Refactor database testing: move testDatabase function to testSupabase.js and update imports in router.js
